:wrench: Rename TopBar file to index.
